requirements.txt
aiofiles==0.3.2
aiohttp==3.1.3
aiohttp-jinja2==1.0.0
aiohttp-theme==0.1.4
aiojobs==0.2.1
aiopg==0.13.2
async-timeout==2.0.1
asyncio==3.4.3
Jinja2==2.10
multidict==4.2.0
psycopg2==2.7.3.1
psycopg2-binary==2.7.4
python-dateutil==2.7.3
PyYAML==3.12
SQLAlchemy==1.2.7
sqlparse==0.2.4
Unidecode==0.4.21
urllib3==1.22
uvloop==0.9.1
yarl==1.2.3
